fix(ecr-build): add random sleep before fetching AWS credentials
There's currently an issue with the GitHub OIDC provider handling
multiple requests at the same time meaning that assuming an AWS role
using OIDC fails if too many requests happen in parallel.

Until the relevant PR is merged to implement retrying in the
configure-aws-credentials action, we can sleep for a random amount of
time before fetching the credentials to reduce the chance of a clash.

See: aws-actions/configure-aws-credentials#350