feat(tf-aws): workflow for running Terraform against AWS
The tf-aws workflow runs Terraform against the AWS accounts in our
organization based on the presence of tfvars files in the target
repository. The workflow uses GitHub+AWS OIDC authentication to restrict
the scope of the IAM roles based on the name of the target repository
and whether the workflow has been triggered during a PR or a push to the
`main` branch.

The workflow expects the target repository (or a subdirectory specified
with the `path` input) to contain a single Terraform "root module" and a
`vars/` directory containing a `.tfvars` file for each AWS account we
want to apply resources from that module to.

Roles used by this workflow are managed in the 3DHubs/infra-github-roles
repository.