feat(tf-aws): add iamlive step
iamlive is a tool that processes the calls made to AWS and automatically
determines a suitable IAM policy. By combining it with localstack, we
can try and "guess" the minimum viable policy for running Terraform in a
given repo.

The step isn't essential for the tf-aws workflow, and is provided only
as a maybe useful tool for developing a suitable repository policy to
be defined in the infra-github-roles repository. It probably won't work
when your Terraform config has lots of data sources, for example.